HACKING: update docs with "git request-pull" info
"git request-pull" is not a well-known command, so add a
mention about it to avoid misleading people into thinking
we use a centralized service.

While we're in the area, put an extra link to the HTTPS
archives, since SMTP-based subscriptions are not
centralization-resistant.

Thanks to Ngan Pham for the feedback.

Link: https://yhbt.net/kgio-public/CAAvYYt6ruhhZcA4OmF=NQZ+-j0RJ0vT5f33tUCrREezo5mUCVQ@mail.gmail.com/